# src/main/java/com/example/book/controller/dto/BooksDto.java
package com.example.book.controller.dto;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class BooksDto {

    private List<BookDto> books;
}