# src/main/java/com/example/book/controller/dto/CreateBookDto.java
package com.example.book.controller.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CreateBookDto {

    @NotBlank
    private String name;

    @NotBlank
    private String author;

    private String count;
}